requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
tabulate>=0.9.0
jupyter_client>=8.0.0
ipykernel>=6.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
